#6 fixes for mavenCentral
