+++ .github/workflows/compilation-check.yml
@@ -19,4 +19,8 @@ jobs:
     - name: Cocoapods install
       run: (cd sample/ios-app && pod install)
     - name: Check library
-      run: ./gradlew build publishToMavenLocal
+      run: ./gradlew build publishToMavenLocal syncMultiPlatformLibraryDebugFrameworkIosX64
+    - name: Install pods with kotlin
+      run: cd sample/ios-app && pod install
+    - name: build ios sample
+      run: cd sample/ios-app && xcodebuild -scheme TestProj -workspace TestProj.xcworkspace -configuration Debug build CODE_SIGNING_REQUIRED=NO CODE_SIGNING_ALLOWED=NO
