+++ build.gradle.kts
@@ -2,6 +2,8 @@
  * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
+import java.util.Base64
+
 plugins {
     plugin(Deps.Plugins.detekt) apply false
 }
@@ -25,13 +27,17 @@ buildscript {
 
 allprojects {
     repositories {
+        mavenCentral()
         google()
-        jcenter()
 
-        maven { url = uri("https://kotlin.bintray.com/kotlin") }
-        maven { url = uri("https://kotlin.bintray.com/kotlinx") }
         maven { url = uri("https://dl.bintray.com/aakira/maven") }
-        maven { url = uri("https://dl.bintray.com/icerockdev/moko") }
+
+        jcenter {
+            content {
+                includeGroup("org.jetbrains.trove4j")
+                includeGroup("org.jetbrains.kotlinx")
+            }
+        }
     }
 
     apply(plugin = Deps.Plugins.detekt.id)
@@ -55,6 +61,75 @@ allprojects {
         }
     }
 
+    plugins.withId(Deps.Plugins.mavenPublish.id) {
+        group = "dev.icerock.moko"
+        version = Deps.mokoCrashReportingVersion
+
+        val javadocJar by tasks.registering(Jar::class) {
+            archiveClassifier.set("javadoc")
+        }
+
+        configure<PublishingExtension> {
+            repositories.maven("https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/") {
+                name = "OSSRH"
+
+                credentials {
+                    username = System.getenv("OSSRH_USER")
+                    password = System.getenv("OSSRH_KEY")
+                }
+            }
+
+            publications.withType<MavenPublication> {
+                // Stub javadoc.jar artifact
+                artifact(javadocJar.get())
+
+                // Provide artifacts information requited by Maven Central
+                pom {
+                    name.set("MOKO crash reporting")
+                    description.set("Fatal and Non-Fatal reporting to Crashlytics for Kotlin Multiplatform Mobile")
+                    url.set("https://github.com/icerockdev/moko-crash-reporting")
+                    licenses {
+                        license {
+                            url.set("https://github.com/icerockdev/moko-crash-reporting/blob/master/LICENSE.md")
+                        }
+                    }
+
+                    developers {
+                        developer {
+                            id.set("Dorofeev")
+                            name.set("Andrey Dorofeev")
+                            email.set("adorofeev@icerockdev.com")
+                        }
+                        developer {
+                            id.set("Alex009")
+                            name.set("Aleksey Mikhailov")
+                            email.set("aleksey.mikhailov@icerockdev.com")
+                        }
+                    }
+
+                    scm {
+                        connection.set("scm:git:ssh://github.com/icerockdev/moko-crash-reporting.git")
+                        developerConnection.set("scm:git:ssh://github.com/icerockdev/moko-crash-reporting.git")
+                        url.set("https://github.com/icerockdev/moko-crash-reporting")
+                    }
+                }
+            }
+
+            apply(plugin = Deps.Plugins.signing.id)
+
+            configure<SigningExtension> {
+                val signingKeyId: String? = System.getenv("SIGNING_KEY_ID")
+                val signingPassword: String? = System.getenv("SIGNING_PASSWORD")
+                val signingKey: String? = System.getenv("SIGNING_KEY")?.let { base64Key ->
+                    String(Base64.getDecoder().decode(base64Key))
+                }
+                if (signingKeyId != null) {
+                    useInMemoryPgpKeys(signingKeyId, signingKey, signingPassword)
+                    sign(publications)
+                }
+            }
+        }
+    }
 }
 
 tasks.register("clean", Delete::class).configure {
