+++ crash-reporting-core/build.gradle.kts
@@ -2,78 +2,13 @@
  * Copyright 2020 IceRock MAG Inc. Use of this source code is governed by the Apache 2.0 license.
  */
 
-import java.util.Base64
-import kotlin.text.String
-
 plugins {
     plugin(Deps.Plugins.androidLibrary)
     plugin(Deps.Plugins.kotlinMultiplatform)
     plugin(Deps.Plugins.mobileMultiplatform)
     plugin(Deps.Plugins.mavenPublish)
-    plugin(Deps.Plugins.signing)
 }
-
-group = "dev.icerock.moko"
-version = Deps.mokoCrashReportingVersion
 
 dependencies {
     androidMainImplementation(Deps.Libs.Android.appCompat)
-}
-
-val javadocJar by tasks.registering(Jar::class) {
-    archiveClassifier.set("javadoc")
-}
-
-publishing {
-    repositories.maven("https://s01.oss.sonatype.org/service/local/staging/deploy/maven2/") {
-        name = "OSSRH"
-
-        credentials {
-            username = System.getenv("OSSRH_USER")
-            password = System.getenv("OSSRH_KEY")
-        }
-    }
-
-    publications.withType<MavenPublication> {
-        // Stub javadoc.jar artifact
-        artifact(javadocJar.get())
-
-        // Provide artifacts information requited by Maven Central
-        pom {
-            name.set("MOKO crash reporting")
-            description.set("Fatal and Non-Fatal reporting to Crashlytics for Kotlin Multiplatform Mobile")
-            url.set("https://github.com/icerockdev/moko-crash-reporting")
-            licenses {
-                license {
-                    url.set("https://github.com/icerockdev/moko-crash-reporting/blob/master/LICENSE.md")
-                }
-            }
-
-            developers {
-                developer {
-                    id.set("Alex009")
-                    name.set("Aleksey Mikhailov")
-                    email.set("aleksey.mikhailov@icerockdev.com")
-                }
-            }
-
-            scm {
-                connection.set("scm:git:ssh://github.com/icerockdev/moko-crash-reporting.git")
-                developerConnection.set("scm:git:ssh://github.com/icerockdev/moko-crash-reporting.git")
-                url.set("https://github.com/icerockdev/moko-crash-reporting")
-            }
-        }
-    }
-
-    signing {
-        val signingKeyId: String? = System.getenv("SIGNING_KEY_ID")
-        val signingPassword: String? = System.getenv("SIGNING_PASSWORD")
-        val signingKey: String? = System.getenv("SIGNING_KEY")?.let { base64Key ->
-            String(Base64.getDecoder().decode(base64Key))
-        }
-        if (signingKeyId != null) {
-            useInMemoryPgpKeys(signingKeyId, signingKey, signingPassword)
-            sign(publishing.publications)
-        }
-    }
 }
