+++ sample/ios-app/src/GoogleService-Info.plist
@@ -1,0 +1,36 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
+<plist version="1.0">
+<dict>
+	<key>CLIENT_ID</key>
+	<string>149963383979-cdc8ptbcfqdofkk5jfdm4ln8m7ngv6p8.apps.googleusercontent.com</string>
+	<key>REVERSED_CLIENT_ID</key>
+	<string>com.googleusercontent.apps.149963383979-cdc8ptbcfqdofkk5jfdm4ln8m7ngv6p8</string>
+	<key>API_KEY</key>
+	<string>AIzaSyBnehCN7AvIrtno7WGpCjyHfY3LQZ6f-3w</string>
+	<key>GCM_SENDER_ID</key>
+	<string>149963383979</string>
+	<key>PLIST_VERSION</key>
+	<string>1</string>
+	<key>BUNDLE_ID</key>
+	<string>dev.icerock.moko.sample.crashReporting-test</string>
+	<key>PROJECT_ID</key>
+	<string>moko-crash-report</string>
+	<key>STORAGE_BUCKET</key>
+	<string>moko-crash-report.appspot.com</string>
+	<key>IS_ADS_ENABLED</key>
+	<false></false>
+	<key>IS_ANALYTICS_ENABLED</key>
+	<false></false>
+	<key>IS_APPINVITE_ENABLED</key>
+	<true></true>
+	<key>IS_GCM_ENABLED</key>
+	<true></true>
+	<key>IS_SIGNIN_ENABLED</key>
+	<true></true>
+	<key>GOOGLE_APP_ID</key>
+	<string>1:149963383979:ios:70426b87b87dee8ae543ee</string>
+	<key>DATABASE_URL</key>
+	<string>https://moko-crash-report.firebaseio.com</string>
+</dict>
+</plist>
